Create PS-Script
We had an issue where the UPN of our users was still set from when we had Windows 2000 AD's and contained spaces etc. The guy that did the Identity Project in 2005 seemed to think that was ok so all users are provisioned like that even now. In the short term until we rebuild this renames all UPN's to the mail nickname.